+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Geode Changelog
 
+## v5.0.0-beta.4
+ * Add `requires-patching` mod.json key for iOS (3b5e715)
+ * Add `geode::Button` (#1810)
+ * Update network libraries (curl, openssl, etc)
+ * Fix recursion and race conditions on Events (#1808, #1817)
+ * Fix `CCLabelBMFont` ctor crash (96bab68)
+ * Add iOS controller support, and fix some input bugs (#1742)
+ * Improve mod dependency popup (#1797)
+ * Make `async::waitForMainThread` fully cancel safe (4b851bd)
+ * Properly implement DNS lookup on Android (73d61b9)
+ * Fix `NineSlice` opacity/color cascade behavior (#1805)
+
 ## v5.0.0-beta.3
  * Add keybind categories, and a global list of keybinds (134878e, b16c2b9, d4eb55a, 3bfe471, 9de1bf1, db160d8)
  * Keybind listener improvements (#1782)